Enable eq role by default
