+++ doc/extdev/index.rst
@@ -146,6 +146,11 @@ The following is a list of deprecated interface.
      - 3.0
      - ``docutils.nodes.math_block``
 
+   * - ``sphinx.ext.mathbase.eqref`` (node)
+     - 1.8
+     - 3.0
+     - ``sphinx.addnodes.math_reference``
+
    * - ``viewcode_import`` (config value)
      - 1.8
      - 3.0
