+++ sphinx/config.py
@@ -140,6 +140,8 @@ class Config(object):
         numfig_format = ({}, 'env', []),  # will be initialized in init_numfig_format()
 
         math_number_all = (False, 'env', []),
+        math_eqref_format = (None, 'env', string_classes),
+        math_numfig = (True, 'env', []),
         tls_verify = (True, 'env', []),
         tls_cacerts = (None, 'env', []),
         smartquotes = (True, 'env', []),
